Sistemas de Aprendizaje Automaticos/Aprendizaje_Supervisado/requirements.txt
flask==3.1.0
scikit-learn==1.6.1
numpy==2.2.3
Pillow==11.1.0
